Renamed OpportunityCardHeader
