+++ apps/charterafrica/src/components/OpportunityCards/OpportunityCardHeader.snap.js
